LayoutTests/editing/deleting/merge-paragraph-from-span-with-multiple-text-decoration-live-range.html

<!DOCTYPE html><!-- webkit-test-runner [ LiveRangeSelectionEnabled=true ] -->
<html>
<head>
<style type="text/css">
.MultipleTextDecorations {
text-decoration: underline overline line-through;
}
</style>
</head>
<body>
<div id="test" contenteditable="true">
<p>foo</p>
<span id="spanToMerge" class="MultipleTextDecorations">bar</span>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
Markup.description('The span style should be preserved when merging a pragraph.\n'
    + 'The test passes if the "bar" has underline, overline and line through.');

var spanToMerge = document.getElementById("spanToMerge");
spanToMerge.focus();
getSelection().collapse(spanToMerge, 0);
document.execCommand("Delete");
Markup.dump(document.getElementById("test"));
</script>
 </body>
</html>
