LayoutTests/editing/deleting/merge-paragraph-from-span-with-multiple-text-decoration-live-range-expected.txt

The span style should be preserved when merging a pragraph.
The test passes if the "bar" has underline, overline and line through.
| "\n"
| <p>
|   "foo<#selection-caret>"
|   <span>
|     style="text-decoration: underline overline line-through;"
|     "bar"
